% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{length.exstra_score}
\alias{length.exstra_score}
\title{Number of data points in exstra_score object}
\usage{
\method{length}{exstra_score}(x)
}
\arguments{
\item{x}{exstra_score object.}
}
\value{
A numeric of length 1
}
\description{
Number of data points in exstra_score object
}
\examples{
data(exstra_wgs_pcr_2)
length(exstra_wgs_pcr_2)
}
